requirements.txt.txt
fastapi
uvicorn
pandas
networkx
python-pptx
pydantic